[Inductor cutlass backend] Generate standalone debug runners for Cutlass GEMMs
For debugging and code validation purposes, it's often helpful to be able to
run the generated GEMM Kernels in a standalone manner, without Python and Pytorch.

This diff adds a bit of code to GEMM Kernels which allows to compile and run
them as standalone executables, easing debugging, profiling and memory-checking
with CUDA Toolkit based tools.

ghstack-source-id: 06f1e093bef53bdea9df66a543a3df18bc0c2171
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115072